CHET-289: refactor MigrationTransfer component into multiple smaller component
